Parametrize for tests

+++ tests/image/test_allocate.py
@@ -1,15 +1,26 @@
 import pytest
 
-from api            import One
-from pyone          import OneServer, OneActionException, OneNoExistsException, OneException
-from utils          import get_user_auth, run_command
-from one_cli.image  import Image, image_exist
-from config         import API_URI, COMMAND_EXECUTOR, BRESTADM
+from api                import One
+from pyone              import OneActionException, OneNoExistsException, OneException
+from utils              import get_user_auth, create_temp_file, delete_temp_file
+from one_cli.image      import Image, image_exist
+from one_cli.datastore  import Datastore
+from config             import BRESTADM
 
 
 
-BRESTADM_AUTH     = get_user_auth(BRESTADM)
-BRESTADM_SESSION  = OneServer(API_URI, BRESTADM_AUTH)
+BRESTADM_AUTH          = get_user_auth(BRESTADM)
+STORAGE_IMAGE_TEMPLATE = """
+NAME   = test_api_img_storage
+TYPE   = IMAGE_DS
+TM_MAD = ssh
+DS_MAD = fs
+"""
+STORAGE_SYSTEM_TEMPLATE = """
+NAME   = test_api_sys_storage
+TYPE   = SYSTEM_DS
+TM_MAD = ssh
+"""
 
 
 
@@ -18,90 +29,63 @@ BRESTADM_SESSION  = OneServer(API_URI, BRESTADM_AUTH)
 # TESTS
 # =================================================================================================
 
-
-def test_create_datablock_and_dont_check_capacity():
+@pytest.mark.parametrize("datastore", [STORAGE_IMAGE_TEMPLATE,], indirect=True)
+@pytest.mark.parametrize("one", [BRESTADM_AUTH,], indirect=True)
+@pytest.mark.parametrize("cap_check", [True, False])
+def test_capacity_check(one: One, datastore: Datastore, cap_check: bool):
     template = """
         NAME = api_test_image_1
         TYPE = DATABLOCK
         SIZE = 1024000
     """
-    one       = One(BRESTADM_SESSION)
-    image_id  = one.image.allocate(template, 1, False)
-    image     = Image(image_id)
+    if cap_check:
+        with pytest.raises(OneActionException):
+            one.image.allocate(template, datastore._id, cap_check)
+    else:
+        image_id  = one.image.allocate(template, datastore._id, cap_check)
+        image     = Image(image_id)
+        assert image_exist(image_id) == True
+        image.delete()
 
-    assert image_exist(image_id) == True
-    image.delete()
 
 
-
-def test_create_datablock_and_check_capacity():
-    template = """
-        NAME = api_test_image_2
-        TYPE = DATABLOCK
-        SIZE = 1024000
-    """
-
-    one  = One(BRESTADM_SESSION)
-    with pytest.raises(OneActionException):
-        one.image.allocate(template, 1, True)
-    
-    
-
-def test_create_datablock_when_datastore_not_exist():
+@pytest.mark.parametrize("one", [BRESTADM_AUTH,], indirect=True)
+def test_datastore_not_exist(one: One):
     template = """
         NAME = api_test_image_3
         TYPE = DATABLOCK
-        SIZE = 1024000 
+        SIZE = 1 
     """
-
-    one  = One(BRESTADM_SESSION)
     with pytest.raises(OneNoExistsException):
         one.image.allocate(template, 99999)
 
 
 
-def test_create_datablock_in_system_datastore():
+@pytest.mark.parametrize("datastore", [STORAGE_SYSTEM_TEMPLATE,], indirect=True)
+@pytest.mark.parametrize("one", [BRESTADM_AUTH,], indirect=True)
+def test_wrong_datastore(one: One, datastore: Datastore):
     template = """
         NAME = api_test_image_4
         TYPE = DATABLOCK
-        SIZE = 10 
+        SIZE = 1
     """
-
-    one  = One(BRESTADM_SESSION)
     with pytest.raises(OneException):
-        one.image.allocate(template, 0)
+        one.image.allocate(template, datastore._id)
 
 
 
-def test_create_datablock_as_persistent():
-    template = """
-        NAME = api_test_image_5
-        TYPE = DATABLOCK
-        SIZE = 100
-        PERSISTENT = YES 
-    """
-    one      = One(BRESTADM_SESSION)
-    image_id = one.image.allocate(template, 1)
-    image    = Image(image_id)
-
-    assert image.info().PERSISTENT == True
-    image.delete()
-
-
-
-def test_create_os_image_by_local_path():
+@pytest.mark.parametrize("datastore", [STORAGE_IMAGE_TEMPLATE,], indirect=True)
+@pytest.mark.parametrize("one", [BRESTADM_AUTH,], indirect=True)
+def test_local_file_load(one: One, datastore: Datastore):
     file_path = "/tmp/testfile"
-    run_command(COMMAND_EXECUTOR + " " + f"dd if=/dev/urandom of={file_path} bs=1MiB count=10")
+    create_temp_file(1, file_path)
     template  = f"""
         NAME = api_test_image_6
         TYPE = OS
         PATH = {file_path}
     """
-    one      = One(BRESTADM_SESSION)
-    image_id = one.image.allocate(template, 1)
+    image_id = one.image.allocate(template, datastore._id)
     image    = Image(image_id)
-
     assert image_exist(image_id) == True
     image.delete()
-
-    run_command(COMMAND_EXECUTOR + " " + f"rm -f {file_path}")
+    delete_temp_file(file_path)
